I promised I'll import Abakus into KDE's Subversion repo, so here it is (attempt 2).
Beware though, it uses bksys, which means it pretty much ignores the rest of
the KDE build system altogether.  I'm actually torn about this:  bksys/SCons is
really frickin' nice, but then I can't automate the build with kdesvn-build.

I've investigated adding Makefile.am and configure.in.in and forwarding the
call to Abakus's configure script and SCons, but it seems to be more trouble
than it's worth.  Oh well.

svn path=/trunk/playground/utils/abakus/; revision=449908